imageModal move Ui folder
